Update BSP for v1.1, add more debug commands
Closes #43
