+++ fw/rbcx-coprocessor/include/ButtonController.hpp
@@ -1,3 +1,4 @@
 #pragma once
 
 void buttonControllerPoll();
+void buttonControllerSetDebug(bool debug);
