+++ fw/rbcx-coprocessor/platformio.ini
@@ -8,21 +8,20 @@
 ; Please visit documentation for the other options and examples
 ; https://docs.platformio.org/page/projectconf.html
 
-[env:rbcx]
+[platformio]
+default_envs = hw11
 
+[env]
 platform = ststm32@~7.0.0
 board = genericSTM32F103VC
 framework = stm32cube
 lib_deps =
     https://github.com/RoboticsBrno/RB3204-RBCX-coproc-comm/archive/2a8cd0b31d3fccf603724f24e5aaba698d30e47a.zip
 
-build_type = debug
 build_flags =
     -Iinclude
     -std=c++17
     -DUSE_FULL_LL_DRIVER
-    -DUSBD_DP_PORT=GPIOA
-    -DUSBD_DP_PIN=8
     -DUSB_PMASIZE=512
     -DRBCX_VER_REVISION='"ffffffff"'
     -DRBCX_VER_DIRTY=1
@@ -36,3 +35,19 @@ extra_scripts =
 
 debug_tool = stlink
 upload_protocol = stlink
+
+[env:hw10]
+build_type = debug
+build_flags =
+    ${env.build_flags}
+    -DRBCX_HW_VER=0x0100
+    -DUSBD_DP_PORT=GPIOA
+    -DUSBD_DP_PIN=8
+
+[env:hw11]
+build_type = debug
+build_flags =
+    ${env.build_flags}
+    -DRBCX_HW_VER=0x0101
+    -DUSBD_DP_PORT=GPIOD
+    -DUSBD_DP_PIN=2
