+++ fw/rbcx-coprocessor/src/Bsp.cpp
@@ -11,9 +11,12 @@
 
 static TaskWrapper<1024> softResetTask;
 
+#if RBCX_HW_VER == 0x0100
 extern "C" void EXTI1_IRQHandler(void) { HAL_GPIO_EXTI_IRQHandler(GPIO_PIN_1); }
-
 extern "C" void EXTI3_IRQHandler(void) { HAL_GPIO_EXTI_IRQHandler(GPIO_PIN_3); }
+#elif RBCX_HW_VER == 0x0101
+extern "C" void EXTI4_IRQHandler(void) { HAL_GPIO_EXTI_IRQHandler(GPIO_PIN_4); }
+#endif
 
 extern "C" void EXTI9_5_IRQHandler(void) {
     HAL_GPIO_EXTI_IRQHandler(GPIO_PIN_5);
@@ -26,9 +29,14 @@ extern "C" void EXTI9_5_IRQHandler(void) {
 extern "C" void HAL_GPIO_EXTI_Callback(uint16_t pin) {
     if (pin == espEnPin.second) {
         sEsp32Manager.onEnRisingInIrq();
+        return;
     }
-    if (pin & utsEchoPins.second) {
-        ultrasoundOnEchoEdge();
+
+    for (const auto& p : utsEchoPin) {
+        if (pin == p.second) {
+            ultrasoundOnEchoEdge();
+            return;
+        }
     }
 }
