+++ fw/rbcx-coprocessor/src/ButtonController.cpp
@@ -4,11 +4,17 @@
 #include "task.h"
 
 #include "Bsp.hpp"
+#include "DebugLink.hpp"
 #include "Dispatcher.hpp"
 #include "Power.hpp"
+#include "utils/Debug.hpp"
 
 static uint32_t buttonState = 0;
 static uint32_t powerOffAt = 0;
+
+static bool debugStateChanges = false;
+
+void buttonControllerSetDebug(bool debug) { debugStateChanges = debug; }
 
 void buttonControllerPoll() {
     if (powerOffAt != 0 && xTaskGetTickCount() >= powerOffAt) {
@@ -19,6 +25,11 @@ void buttonControllerPoll() {
     const uint32_t newButtonState = getButtons();
     if (buttonState == newButtonState)
         return;
+
+    if (debugStateChanges) {
+        DEBUG("button state change: 0x%04lx -> 0x%04lx\n", buttonState,
+            newButtonState);
+    }
 
     auto status = CoprocStat();
     status.which_payload = CoprocStat_buttonsStat_tag;
